mutualize search bar component
